+++ src/changelog.ts
@@ -1,9 +1,9 @@
-import { getOctokit, context } from '@actions/github';
-import * as core from '@actions/core';
+import {getOctokit, context} from '@actions/github'
+import * as core from '@actions/core'
 
-import { IChangeLog } from './interfaces'
-import updateChangelog from "./useCases/updateChangelog"
-import createNewRelease from "./useCases/createNewRelease"
+import {IChangeLog} from './interfaces'
+import updateChangelog from './useCases/update-changelog'
+import createNewRelease from './useCases/create-new-release'
 
 export default async function changelog({
   changelogFileName,
@@ -23,13 +23,11 @@ export default async function changelog({
       encoding
     })
 
-    const branch = core.getInput('branch');
-    const sha = core.getInput('sha');
+    const sha = core.getInput('sha')
 
     await createNewRelease({
       getOctokit,
       context,
-      branch,
       sha,
       changelogFileName,
       encoding
