+++ src/interfaces/create-new-release.ts
@@ -1,9 +1,8 @@
-import { Context } from '@actions/github/lib/context';
+import {Context} from '@actions/github/lib/context'
 
 export default interface ICreateNewRelease {
   getOctokit: any
   context: Context
-  branch: string
   sha?: string
   changelogFileName: string
   encoding: any
