+++ src/interfaces/index.ts
@@ -1,6 +1,6 @@
 import IChangeLog from './change-log'
+import ICreateNewRelease from './create-new-release'
 import IGetOldLogs from './get-old-logs'
 import IMountLog from './mount-log'
-import ICreateNewRelease from './create-new-release'
 
-export { IChangeLog, IGetOldLogs, IMountLog, ICreateNewRelease }
+export {IChangeLog, IGetOldLogs, IMountLog, ICreateNewRelease}
